Rename ReactTypeOfWork to ReactWorkTags
And `type TypeOfWork` to `type WorkTag`.
