+++ packages/events/__tests__/ResponderEventPlugin-test.internal.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const {HostComponent} = require('shared/ReactTypeOfWork');
+const {HostComponent} = require('shared/ReactWorkTags');
 
 let EventPluginHub;
 let ResponderEventPlugin;
