+++ packages/react-reconciler/src/ReactChildFiber.js
@@ -27,7 +27,7 @@ import {
   HostText,
   HostPortal,
   Fragment,
-} from 'shared/ReactTypeOfWork';
+} from 'shared/ReactWorkTags';
 import invariant from 'shared/invariant';
 import warning from 'shared/warning';
 import warningWithoutStack from 'shared/warningWithoutStack';
