+++ packages/react-reconciler/src/ReactFiberReconciler.js
@@ -27,7 +27,7 @@ import {
   HostComponent,
   ClassComponent,
   ClassComponentLazy,
-} from 'shared/ReactTypeOfWork';
+} from 'shared/ReactWorkTags';
 import getComponentName from 'shared/getComponentName';
 import invariant from 'shared/invariant';
 import warningWithoutStack from 'shared/warningWithoutStack';
